Cleanup SDL includes

+++ prboom2/src/i_main.c
@@ -59,6 +59,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+#include <SDL_timer.h>
+
 #include "e6y.h"
 
 /* Most of the following has been rewritten by Lee Killough
@@ -84,8 +86,8 @@ int ms_to_next_tick;
 
 int I_GetTime_RealTime (void)
 {
-  int64_t t = I_GetTime_MS();
-  int64_t i = t * TICRATE / 1000;
+  long long t = I_GetTime_MS();
+  long long i = t * TICRATE / 1000;
 
   ms_to_next_tick = (i + 1) * 1000 / TICRATE - t;
   ms_to_next_tick = BETWEEN(0, 1000 / TICRATE, ms_to_next_tick);
@@ -97,8 +99,8 @@ int realtic_clock_rate = 100;
 
 static int I_GetTime_Scaled(void)
 {
-  int64_t t = I_GetTime_MS();
-  int64_t i = t * TICRATE * realtic_clock_rate / 100000;
+  long long t = I_GetTime_MS();
+  long long i = t * TICRATE * realtic_clock_rate / 100000;
 
   ms_to_next_tick = (i + 1) * 100000 / realtic_clock_rate / TICRATE - t;
   ms_to_next_tick = BETWEEN(0, 100000 / realtic_clock_rate / TICRATE, ms_to_next_tick);
@@ -131,12 +133,12 @@ static int I_TickElapsedTime_FastDemo(void)
 
 static int I_TickElapsedTime_RealTime(void)
 {
-  return (int64_t)I_GetTime_MS() * TICRATE % 1000 * FRACUNIT / 1000;
+  return (long long)I_GetTime_MS() * TICRATE % 1000 * FRACUNIT / 1000;
 }
 
 static int I_TickElapsedTime_Scaled(void)
 {
-  return (int64_t)I_GetTime_MS() * realtic_clock_rate * TICRATE / 100 % 1000 * FRACUNIT / 1000;
+  return (long long)I_GetTime_MS() * realtic_clock_rate * TICRATE / 100 % 1000 * FRACUNIT / 1000;
 }
 
 int (*I_TickElapsedTime)(void) = I_TickElapsedTime_RealTime;
